pyproject.toml
[project]
name = "fastapi-ai-starter"
version = "0.0.1"
description = "A fastapi template"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.1",
    "asgiref>=3.8.1",
    "celery-stubs==0.1.3",
    "celery[redis]>=5.5.2",
    "fastapi-pagination>=0.13.2",
    "fastapi[standard]>=0.115.12",
    "greenlet>=3.2.3",
    "loguru>=0.7.3",
    "passlib[bcrypt]>=1.7.4",
    "pre-commit>=4.2.0",
    "aiomysql>=0.2.0",
    "pydantic>=2.11.4",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "sqlmodel>=0.0.24",
]

[dependency-groups]
dev = [
    "mypy>=1.16.1",
    "pre-commit>=4.2.0",
    "ruff>=0.12.0",
    "types-passlib>=1.7.7.20250602",
]

[tool.mypy]
plugins = ["pydantic.mypy"]
